Rename decompressors classes
