+++ src/ImageSharp/Formats/Tiff/Compression/Compressors/TiffLzwEncoder.cs
