+++ src/ImageSharp/Formats/Tiff/Compression/Decompressors/T4BitReader.cs
@@ -8,7 +8,7 @@ using System.IO;
 
 using SixLabors.ImageSharp.Memory;
 
-namespace SixLabors.ImageSharp.Formats.Experimental.Tiff.Compression
+namespace SixLabors.ImageSharp.Formats.Experimental.Tiff.Compression.Decompressors
 {
     /// <summary>
     /// Bitreader for reading compressed CCITT T4 1D data.
