+++ src/ImageSharp/Formats/Tiff/Compression/Decompressors/TiffBaseCompression.cs
@@ -8,7 +8,7 @@ using SixLabors.ImageSharp.Formats.Experimental.Tiff.Constants;
 using SixLabors.ImageSharp.IO;
 using SixLabors.ImageSharp.Memory;
 
-namespace SixLabors.ImageSharp.Formats.Experimental.Tiff.Compression
+namespace SixLabors.ImageSharp.Formats.Experimental.Tiff.Compression.Decompressors
 {
     /// <summary>
     /// Base tiff decompressor class.
